Renaming files and folders to snake_case
