+++ package.json
@@ -19,7 +19,8 @@
     "lucide-react": "^0.562.0",
     "next": "16.1.3",
     "react": "19.2.3",
-    "react-dom": "19.2.3"
+    "react-dom": "19.2.3",
+    "react-usestateref": "^2.0.0"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.0.0",
