+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import HeroSection from '@/components/home/hero-section';
-import ChainList from '@/components/home/chain-list';
-import JourneySection from '@/components/home/journey-section';
+import HeroSection from '@/components/home/hero_section';
+import ChainList from '@/components/home/chain_list';
+import JourneySection from '@/components/home/journey_section';
 
 export default function Home() {
   return (
