+++ src/components/home/hero_section.tsx
