+++ src/components/home/journey_section.tsx
